File Cleanup
Got rid of old DataMaker file, renamed idontknow to DataMaker since that's what it does
